Simplify connection logic
Relationship is directed.
